Move new classes into separate cs files
